Couple of fixes
1. We need to make sure if we are already closing/draining we shouldn't
close the connection. This is critical because during reconfig we
could decide to close the connection but in the meanwhile we could
have received a drain command and started draining. In that case we
should just allow the drain process to take care of the rest.
2. Instead of waiting for the entire timeout period, we can ideally just
check for the number of responses we received and bail out immediately.