Backoff retry is not needed for ReadConsumerGroupHosts() (#38)
Some users complained that Consumer.Open() does not respect the timeout setting in the client config. After digging I found that ReadConsumerGroupHosts() has a backoff retry wrapping the thrift call. For Open() such retry is not needed. The other use of ReadConsumerGroupHosts() is in reconfiguration which itself is a big retry loop. Therefore, we don't need to have backoff retry for ReadConsumerGroupHosts().